Further work on Inputs. Amount (incomplete).
We now remove the escape characters from the input fields (if they exist).
Next step is to review form of the amount input.
We would also want to handle the input at the input stage.  At the moment
we pass the data to the onward function which expects the input data to
be a float.  This should instead confirm the input data is correctly formed
or allow it to be passed back to the user for correction prior to progression
within the program.